requirements.txt
pytest
flake8
python-dotenv